refactor: moved pump search query provider to its own model
