+++ app/lib/src/features/pumps/screens/pump_list/pumps_list_screen.dart
@@ -5,7 +5,7 @@ import 'package:go_router/go_router.dart';
 import 'package:irrigazione_iot/src/config/routes/routes_enums.dart';
 import 'package:irrigazione_iot/src/features/pumps/data/pump_repository.dart';
 import 'package:irrigazione_iot/src/features/pumps/screens/pump_list/dismiss_pump_controller.dart';
-import 'package:irrigazione_iot/src/features/pumps/screens/pump_list/pump_search_query_result.dart';
+import 'package:irrigazione_iot/src/features/pumps/providers/pump_search_query_result.dart';
 import 'package:irrigazione_iot/src/features/pumps/screens/pump_list/pump_status_controller.dart';
 import 'package:irrigazione_iot/src/features/pumps/widgets/empty_pump_widget.dart';
 import 'package:irrigazione_iot/src/features/pumps/widgets/pump_list_tile.dart';
